Move tests module into a separate folder
